Moved variable to modules
