+++ src/__init__.py
@@ -1,0 +1,7 @@
+#!/bin/env python3
+# -*- coding: utf-8 -*-
+from .import variable   # noqa: F401
+
+__all__ = [
+    'variable'
+]
